+++ CHANGELOG.md
@@ -1,12 +1,20 @@
+## 0.1.4-dev.1
+
+-   Rename `Database` to `Connection`.
+-   Rename `Model` to `Client`.
+-   Add `lastBody` to `Client`.
+-   Complete documentation.
+
 ## 0.1.3-dev.1
 
 -   Added example with testing.
 -   Improve documentation.
--   Added CHANGELOG_PLAN.md
+-   Added CHANGELOG_PLAN.md.
 
 ## 0.1.2-dev.1
 
--   Completed Database, Model, Query system.
+-   Completed `Database`, `Model` system.
+-   Framed in `Query`.
 
 ## 0.1.1-dev.1